test_serial_writer: was failing because double eol
This tester was failing because we were creating SerialWriter instances
with eol='\n' (the default), but then in the write routines we were adding
and extra '\n' manually.  So we had double new-lines, and every other read
was producing a blank line.

Also fixed a typo in a logging.info() call.