Reader/Writer: _encode_str() should not unescape by default
I don't really see any case where we would *want* to unescape during an
encode other than the "probably escaped because we read it from a config
file" case... and that's really a special case (e.g., `eol`, `cmd` params
to PolledSerialReader).  Unescaping data during encode could silently
corrupt outgoing data if unintended, so I'd say the default behavior should
be NOT to unescape.

This commit makes that the case and fixes the small ripple in the code
base.