test_serial_writer: make failures get noticed by unittest
This was originally written with separate threads for readers and writers,
but the reader does a comparison and then calls self.assertTrue(), which
does cause an exception to get raised, but it's not noticed by the
TestCase instance.  The result is you get a traceback on the screen when
you run the tester, but it actually says it was a success at the end.

I'm assuming this is because the thread's target_func doesn't really have
a `self`.   I'm actually surprised it lets you use `self` from within a
non-member function w/out throwing an exception of some sort.

This commit fixes the problem by making the _run_reader() and
_run_read_specialcase() functions into actual member functions, and then
runs them in the foreground.  It also puts our yucky looking special string
in a global variable instead of copy/pasting it in multiple spots.