Fixes for SerialWriter and SerialReader (#346)
* UDPwriter: address review comments from PR #345

- corrected docsting for UDPWriter to just say "records" instead of "text"
- removed `input_format` usage

* SerialWriter: honor the `encoding_error` parameter

The existing code was ignoring the `encoding_error` parameter, which
defaults to 'ignore', and calling str.encode() w/out specifying a failure
mode, which results in 'strict' behavior (opposite of what the docstring
says).  This commit passes the encoding parameters into the base Writer
class and uses its _encode_str() method.

* Reader/Writer: _encode_str() should not unescape by default

I don't really see any case where we would *want* to unescape during an
encode other than the "probably escaped because we read it from a config
file" case... and that's really a special case (e.g., `eol`, `cmd` params
to PolledSerialReader).  Unescaping data during encode could silently
corrupt outgoing data if unintended, so I'd say the default behavior should
be NOT to unescape.

This commit makes that the case and fixes the small ripple in the code
base.

* test_serial_writer: was failing because double eol

This tester was failing because we were creating SerialWriter instances
with eol='\n' (the default), but then in the write routines we were adding
and extra '\n' manually.  So we had double new-lines, and every other read
was producing a blank line.

Also fixed a typo in a logging.info() call.

* SerialReader: Serial.read_until() takes `expected` not `terminator`

Not sure if this is something that changed at some point in pyserial or if
it just never got tested... but as-is, you can't actually set `eol` for a
SerialReader w/ getting an exception here.

* SerialWriter: handle escaped `eol` same as in SerialReader

Ok, it's not 100% the same.  We call _unescape_str() in the constructor
instead of _encode_str() because we actually don't want to encode to
binary yet for the SerialWriter.

* test_serial_writer: make failures get noticed by unittest

This was originally written with separate threads for readers and writers,
but the reader does a comparison and then calls self.assertTrue(), which
does cause an exception to get raised, but it's not noticed by the
TestCase instance.  The result is you get a traceback on the screen when
you run the tester, but it actually says it was a success at the end.

I'm assuming this is because the thread's target_func doesn't really have
a `self`.   I'm actually surprised it lets you use `self` from within a
non-member function w/out throwing an exception of some sort.

This commit fixes the problem by making the _run_reader() and
_run_read_specialcase() functions into actual member functions, and then
runs them in the foreground.  It also puts our yucky looking special string
in a global variable instead of copy/pasting it in multiple spots.

* test_serial_writer: use escaped multi-character `eol` to test more

This commit adds a couple extra characters and an escaped \n to `eol` in
the special writer test just to make sure that gets handled properly.

* SerialReader: improve `eol` handling after reading

This was arbitrarily calling rstrip() on the returned record, but that only
explicitly removes trailing whitespace.  We'll want to remove what was
actually configured as `eol` if something else was passed int.  And we
don't want to strip the record returned by read(size=max_bytes) at all.

* test_serial_writer: test sending binary data

This commit adds a testcase for sending binary data and makes a couple tiny
fixes to SerialWriter and SerialReader to get it to pass.

* test_serial_writer: address trivial review comments from PR #346

- Moved info() call to after write() in two places
- Fixed silly comment typo in SerialReader