UDPReader: remove `eol` parameter
The `eol` parameter as described in the docstring, and as implemented,
really doesn't make sense for UDP where each read gives you an entire
datagram.  TCPReader should do something like this for decoding records
inside a stream, but not here.

NOTE: We *do* need to add something to UDPWriter to detect send() failures
      due to a user's record being too large to send, and perhaps fragment
      the record into multiple datagrams.  Once I've done that, I'll come
      back in here and add fragmentation detection to the read() call,
      which will look ironically similar to the eol logic this commit
      removes.