UDPReader: look for FRAGMENT_MARKER during read()
This commit makes read() loop on recv() calls until it gets a datagram that
doesn't end with FRAGMENT_MARKER.  This is the recv-side of the code in
UDPWriter that detects failed write() due to EMSGSIZE (msg too big to fit
in a single packet).

I decided to notify the user via logger.info() calls when fragmentation is
detected, becaue it's probably unintended and undesirable.