test_udp_reader: removed 2nd thread from broadcast and multicast tests
This commit removes the 2nd reader threads from both the broadcast and
multicast tests.  This wasn't actually testing the one-to-many behavior
of broadcast/multicast, as its still running all on a single host.  It was
actually testing what happens when you have two separate processes bind to
the same (addr,port), which normally would be an error but we've got
REUSEPORT set, so it's strange but works but has nothing to do with
broadcast or multicast.