Replace NetworkWriter with TCPWriter (#348)
* NetworkWriter: do hostname resolution once in constructor

This commit attempts to resolve the host portion of `network` once in the
constructor, to catch unresolvable hosts.

* NetworkWriter: don't convert records to JSON

The NetworkWriter class shouldn't be checking types and encoding things as
JSON.  That breaks the whole Reader/Transform/Writer design, preventing
users from being able to send unmodified records out of a socket
(e.g., sending raw/binary data is impossible).

If somebody wants that behavior, they can use a ToJSONTransform

* NetworkWriter: fix retry loop counting

If num_retrys is 0, we never try at all.  Changing this comparison to <=
makes us try once, then retry up to num_retry times, which is the behavior
described in the docstring.

* NetworkWriter: add `encoding` parameter to allow for raw/binary output

This code adds `encoding` and `encoding_errors` and the exact same behavior
already found in NetworkReader: utf-8 by default, empty means just send raw
bytes, ignore encoding errors by default.

* NetworkWriter: unescape `eol` parameter

We do this already in SerialWriter, SerialReader, NetworkReader, and
UDPReader, but need it here, too.

* NetworkWriter: remove redundant UDP code

Instead of making all the same fixes in NetworkWriter that were just
recently made in UDPWriter, this commit removes the UDP portion of
NetworkWriter.  It now just attempts to do TCP.

* NetworkWriter: take `destination` and `port` like UDPWriter

This commit changes `network` parameter into `port` and `destination` to
match the parameters for UDPWriter.

* NetworkWriter: added `warning_limit` parameter to be like UDPWriter

This commit adds `warning_limit` and associated failure counting and muting
logic to be more like UDPWriter.

* test_network_writer: revamp tcp testing

We now have a bunch of TCP test cases: simple text, text w/ eol, text with
multi-char escaped eol, and binary.  And they all pass.  :-)

* writers: rename NetworkWriter -> TCPWriter

This commit renames NetworkWriter into TCPWriter, since that's all it
does now.

To make the change more obvious, this commit also leaves behind an
empty NetworkWriter class that just prints error messages telling
users to use TCPWriter and/or UDPWriter instead.

* TCPWriter: revamp the connection handling code

This commit makes TCP errors cause subsequent write() calls to reconnect
the socket.  Basically, the whole connect->write sequence is attempted up
to 1+`num_retry` times before a write() finally fails.

This commit also makes a couple trivial tweaks to UDPWriter to keep the
behavior of both classes consistent.

* listen.py: usage changes for Network/UDP Writers and Readers

* listen.py: added --write_tcp to use new TCPWriter class

* listen.py: add --encoding so we can handle raw/binary traffic

This commit adds --encoding to the applicable invocations, so we can use
the cli interface for raw/binary traffic.

* UDPWriter,TCPWriter: fix a few comment typos

* TCPWriter: switch order of `destination` and `port`

The original order was to match that of UDPWriter, but it just feels
unnatural to specify port before destination.

* listen.py: fix arg parsing so ':' is optional in UDP/TCP readers

This was the original behavior, which I didn't really mean to change.

In addition, this commit adds notes to the --help output of
--write_network, --write_udp, and --write_tcp explaining that NetworkWriter
has been replaced by UDPWriter and TCPWriter.