UDPWriter: add record fragmentation support
This commit adds code to write() to detect records that are too big to fit
in a single UDP datagram and automatically fragments them into smaller
writes.  There's code in UDPReader to detect when we've done this by
looking for the FRAGMENT_MARKER we append, and it transparently reassembles
the record as if it had been a single datagram.

I notify the user via logger.info() calls when fragmentation is needed,
because it's probably unintended and undesirable.