Make FileWriter support raw/binary output and other fixes (#350)
* Reader,Writer: ensure encoding='' is same as encoding=None

This commit makes sure '' behaves the same as None, which is what all the
docstrings say, and would be logical... but then certain things treat them
differently (e.g., file.open(mode='ab', encoding='') throws
ValueError: binary mode doesn't take an encoding argument)

* FileWriter: add `encoding` parameter to allow for raw/binary output

This code adds `encoding` and `encoding_errors` and the exact same behavior
already found elsewhere: utf-8 by default, empty means just send raw bytes,
ignore encoding errors by default.

The unique twist for FileWriter is that `encoding` and `mode` have to jive.
If you specify 'ab' and any valid encoding, open() will throw a ValueError.
Similar but different, open() treats encoding=None as "use system default",
so open(mode='a', encoding=None) gives you UTF-8 if that was the system
default.

* FileWriter: don't force record into str()

Removed code auto-detecting when record is of type dict and turning it into
a string before writing.  We're only writing text or binary data coming
from a Reader... not instances of python types.

* FileWriter: allow for use of timestamp as filename prefix

Current code creates a nice timestamp suffix based on the `time_format`
parameter.  This commit checks to see if the last character of
`time_format` is either '-' or '.' and instead uses the timestamp as a
prefix if it is.

* FileWriter: make sure `time_format` is a suffix if format is awkward

Original behavior was `time_format` is always a suffix, so let's change the
conditional code in write() so that all awkward cases retain that behavior.
For example, having '-' on both sides of the `time_format` or not having
it on either end, now both result in a suffix.