Remove deprecated `interface` from UDPWriter and a couple misc fixes (#351)
* UDPWriter: remove deprecated `interface` parameter

This old thing has been deprecated since 2020 (prior to 1.0.0 release), and
is making a bit of a mess out of the logic in here.  Nuke it.  I wasn't
going to do this, but since I'm touching parameters to UDPWriter anyways,
now's the time.

* UDPWriter: make docstring more true regarding broadcast `destination`

'<broadcast>' and '255.255.255.255' are the same and neither sends to all
available interfaces.  The system's network stack sends exactly one packet
out of one interface which it picks (poorly, usually, if you have more than
one interface).  This is one of the reasons using INADDR_BROADCAST is
frowned upon nowadays.

* UDPWriter: switch order of `destination` and `port`

It just feels unnatural to specify port before destination, even if dest
is optional.  This commit swaps them.  Behavior is pretty much the same,
you can still specify destination as '' or None to broadcast... you just
*have* to now.

* UDPWriter,TCPWriter: cleaned up some docstrings