listen.py: fixed UDPReader usage
This commit makes construction of UDPReader, UDPWriter, and TCPWriter
consistent in regards to the dest, port arguments.  Now that all 3 classes
take them in the same order, we just pass them in unnamed as the 1st two
parameters.

This commit also removes `eol` from the UDPReader setup, since we removed
that parameter.