UDPReader,UDPWriter: make source/dest IP optional
This commit makes the 1st constructor parameter in both the UDP reader and
writer classes an optional keyword argument.  This was already almost the
case, as there's code in both classes to treat None or '' specially, but
it was still a required argument.  This ensures that users don't have to
specify destination: '' in their config YML files.

In order to do this, we have to also make `port` a kwarg (because that's
just how Python works).  Since `port` is really required, we check to see
if it was provided and throw a TypeError if it's missing.