Revamp UDPReader (#352)
* UDPReader: increase READ_BUFFER_SIZE and remove `read_buffer_size` param

The UDP header's `length` field sets a theoretical limit of 65,535 bytes
(8-byte header + 65,527 bytes of data) for a UDP datagram.  Technically,
IPV4 or IPv6 headers use up some of that size, so actual maximum data sent
per datagram is slightly less.

UDP receivers should always request the max, though, because if you request
less than what's on the wire, you get what you asked for and the rest gets
tossed on the floor.  There's no built-in error detection/correction in
UDP, so that's unrecoverable data loss.

My testing in python shows that you can recv() w/ pretty much any
ridiculously silly large number (655350, aka 10x the max size) on both
Linux and Mac.  We'll have to put some logic into UDPWriter, though,
because send() is where we run into real-life limits (as opposed to
"theoretical" ones).

Given all that, allowing users to configure the read size via the
`read_buffer_size` parameter is a bad idea: it just allows them to break
things.  So this commit also removes that.

* UDPReader: remove output_format=Text

This commit removes the output_format=Text usage because it's being phased
out and it's technically a lie now that we accept raw/binary.

* UDPReader: fixed up multicast group joining and interface selection

This commit adds `interface` to specify the interface to use (None or ''
cause INADDR_ANY), renames the `source` parameter to `mc_group` to be more
obvious and match verbiage in UDPWriter, splits the socket init code out
into a new _open_socket() method, and initializes the socket as needed in
the write() method (matching behavior of UDPWriter and TCPWriter).

* test_udp_reader: revamp and use new UDPReader parameters

This commit addresses all the parameter changes made to the UDPReader and
UDPWriter classes and revamps the tester to use different test functions
for each test.

* UDPReader: remove `eol` parameter

The `eol` parameter as described in the docstring, and as implemented,
really doesn't make sense for UDP where each read gives you an entire
datagram.  TCPReader should do something like this for decoding records
inside a stream, but not here.

NOTE: We *do* need to add something to UDPWriter to detect send() failures
      due to a user's record being too large to send, and perhaps fragment
      the record into multiple datagrams.  Once I've done that, I'll come
      back in here and add fragmentation detection to the read() call,
      which will look ironically similar to the eol logic this commit
      removes.

* test_udp_reader: added unicast and unicast_binary tests

This commit adds a simple unicast test and a unicast_binary test so that
we can make sure raw/binary data can be read properly.  And it is.

* test_udp_reader: removed 2nd thread from broadcast and multicast tests

This commit removes the 2nd reader threads from both the broadcast and
multicast tests.  This wasn't actually testing the one-to-many behavior
of broadcast/multicast, as its still running all on a single host.  It was
actually testing what happens when you have two separate processes bind to
the same (addr,port), which normally would be an error but we've got
REUSEPORT set, so it's strange but works but has nothing to do with
broadcast or multicast.

* UDPReader,UDPWriter,TCPWriter: added `reuseaddr` and `reuseport` parameters

This commit adds `reuseaddr` and `reuseport` parameters to the UDPReader,
UDPWriter, and TCPWriter classes (TCPReader will also have these it once
implemented).  We were enabling both of these unconditionally, this commit
allows users to enable/disable them.

SO_REUSEADDR is used to prevent toublesome "address already in use" errors
on the listen/accept side of TCP sockets (so it should probably be enabled
by default in TCPReader, once implemented).  I'm honestly not sure what (if
any) effect it has on UDP... I didn't think it had anything to do with UDP,
but google does reveal people claiming it has an effect on multicast
traffic.  I've never noticed a need for SO_REUSEADDR for multicast sockets,
though.  Either way, I've set the default to False, because it's the
socket-level default already.

I don't think I've ever wanted to use SO_REUSEPORT on any socket (TCP, UDP,
unicast, multicast, etc), and I can't imagine a use-case for it on a
deployed system, so it's disabled by default.  But, hey, users can still
enable it if they're brave.

* test_udp_reader: TestCase assertions must come from main thread

Already fixed this in test_serial_writer, didn't realize it was happening
elsewhere.  TestCase assertions MUST come from the main thread, otherwise
they don't get handled (you end up with traceback on stdout, but unittest
says all your tests passed).

This commit moves all our reader threads into the main thread.  It also
increments our port numbers so we can tell from error messages which test
case is having network problems.

* listen.py: fixed UDPReader usage

This commit makes construction of UDPReader, UDPWriter, and TCPWriter
consistent in regards to the dest, port arguments.  Now that all 3 classes
take them in the same order, we just pass them in unnamed as the 1st two
parameters.

This commit also removes `eol` from the UDPReader setup, since we removed
that parameter.

* listen.py: set logging level after parsing --verbose

Looks like we had all the parsing and prep work done, but didn't actually
call setLevel().  This commit adds that.

* UDPReader: improve error message on failure to _open_socket()

* listen.py: removed extraneous print statement

* UDPReader,UDPWriter: make source/dest IP optional

This commit makes the 1st constructor parameter in both the UDP reader and
writer classes an optional keyword argument.  This was already almost the
case, as there's code in both classes to treat None or '' specially, but
it was still a required argument.  This ensures that users don't have to
specify destination: '' in their config YML files.

In order to do this, we have to also make `port` a kwarg (because that's
just how Python works).  Since `port` is really required, we check to see
if it was provided and throw a TypeError if it's missing.