Add record fragmentation support to UDPWriter/UDPReader (#355)
* UDPReader,UDPWriter: make source/dest IP optional, v2

Oops, looks like bind() doesn't like None.  I must have only been testing
previously w/ ''.

* UDPReader: look for FRAGMENT_MARKER during read()

This commit makes read() loop on recv() calls until it gets a datagram that
doesn't end with FRAGMENT_MARKER.  This is the recv-side of the code in
UDPWriter that detects failed write() due to EMSGSIZE (msg too big to fit
in a single packet).

I decided to notify the user via logger.info() calls when fragmentation is
detected, becaue it's probably unintended and undesirable.

* UDPWriter: add record fragmentation support

This commit adds code to write() to detect records that are too big to fit
in a single UDP datagram and automatically fragments them into smaller
writes.  There's code in UDPReader to detect when we've done this by
looking for the FRAGMENT_MARKER we append, and it transparently reassembles
the record as if it had been a single datagram.

I notify the user via logger.info() calls when fragmentation is needed,
because it's probably unintended and undesirable.

* test_udp_reader: added fragmentation test

This commit adds a test that writes a big packet (not huge, nowhere near
the actual maxsize) after lowering the autodetected maxsize to 1k.