Replaces subprocess git tag retrieval with GitPython (#483)
* Replaces git tag retrieval with GitPython

Uses GitPython library to retrieve the current git tag. This change addresses potential issues related to relying on shell commands, and improves error handling. Additionally, it adds GitPython as a dependency.

* Replaces GitPython with subprocess for Git info

Switches from using the GitPython library to using subprocess calls to `git` for retrieving Git information.

This change addresses potential installation issues related to GitPython and simplifies the process of obtaining Git metadata such as tag, remote URL, branch, and commit hash. It avoids the need for users to install GitPython separately.

* Renames git_tag.py to git_info.py

Renames the file to better reflect its purpose
and the type of information it provides.